# contracts/atracciones-openapi.yaml
openapi: 3.0.3
info:
  title: API de Atracciones Turísticas
  description: Microservicio encargado de la gestión del catálogo de atracciones turísticas, sus detalles y búsqueda (Booking Demand API Style).
  version: 1.1.0
servers:
  - url: /api/v1
    description: API Gateway - Ruta base
tags:
  - name: Atracciones
    description: Endpoints para gestionar las atracciones
paths:
  /atracciones/search:
    post:
      tags:
        - Atracciones
      summary: Búsqueda de atracciones
      description: Endpoint para buscar atracciones en base a fechas, locaciones y filtros.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAtraccionesRequest'
      responses:
        '200':
          description: Resultados de la búsqueda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAtraccionesResponse'
        '400':
          $ref: '#/components/responses/ProblemDetails400'
  /atracciones/health:
    get:
      tags:
        - Atracciones
      summary: Healthcheck del microservicio
      responses:
        '200':
          description: Servicio de atracciones operativo
  /atracciones:
    get:
      tags:
        - Atracciones
      summary: Obtener el listado paginado de atracciones (CRUD)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Listado de atracciones recuperado exitosamente.
          headers:
            X-API-Deprecation-Date:
              schema:
                type: string
              description: Fecha en la que esta versión de la API será deprecada.
            Cache-Control:
              schema:
                type: string
              description: Directivas de caché.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAtraccionResponse'
    post:
      tags:
        - Atracciones
      summary: Registrar una nueva atracción
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAtraccionRequest'
      responses:
        '201':
          description: La atracción ha sido creada exitosamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtraccionResponse'
        '400':
          $ref: '#/components/responses/ProblemDetails400'
  /atracciones/{id}:
    get:
      tags:
        - Atracciones
      summary: Obtener el detalle de una atracción
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalle de la atracción.
          headers:
            X-API-Deprecation-Date:
              schema:
                type: string
              description: Fecha en la que esta versión de la API será deprecada.
            Cache-Control:
              schema:
                type: string
              description: Directivas de caché.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtraccionResponse'
        '404':
          $ref: '#/components/responses/ProblemDetails404'
  /atracciones/{id}/availability:
    get:
      tags:
        - Atracciones
      summary: Consultar disponibilidad de cupos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Disponibilidad recuperada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '404':
          $ref: '#/components/responses/ProblemDetails404'
  /atracciones/{id}/reservations:
    post:
      tags:
        - Atracciones
      summary: Crear una reserva de la atracción
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '201':
          description: Reserva confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/ProblemDetails400'
        '404':
          $ref: '#/components/responses/ProblemDetails404'
    put:
      tags:
        - Atracciones
      summary: Reemplazar datos de una atracción
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAtraccionRequest'
      responses:
        '204':
          description: Reemplazo exitoso
        '400':
          $ref: '#/components/responses/ProblemDetails400'
        '404':
          $ref: '#/components/responses/ProblemDetails404'
    patch:
      tags:
        - Atracciones
      summary: Actualizar parcialmente una atracción
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAtraccionRequest'
      responses:
        '200':
          description: Actualización exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtraccionResponse'
        '400':
          $ref: '#/components/responses/ProblemDetails400'
        '404':
          $ref: '#/components/responses/ProblemDetails404'
    delete:
      tags:
        - Atracciones
      summary: Eliminar una atracción
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Eliminación exitosa
        '404':
          $ref: '#/components/responses/ProblemDetails404'

components:
  responses:
    ProblemDetails400:
      description: Petición inválida
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ProblemDetails404:
      description: No encontrado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    Price:
      type: object
      properties:
        currency:
          type: string
          example: USD
        total:
          type: number
          example: 45.0
    Coordinates:
      type: object
      properties:
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
    Location:
      type: object
      properties:
        address:
          type: string
        city:
          type: integer
        country:
          type: string
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        type:
          type: string
    Photo:
      type: object
      properties:
        url:
          type: string
          format: uri
    Operator:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: "Quito Tour Bus"
    Operator:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: "Quito Tour Bus"
    Rating:
      type: object
      properties:
        number_of_reviews:
          type: integer
        score:
          type: number
          format: float
    SearchAtraccionesRequest:
      type: object
      properties:
        currency:
          type: string
          example: EUR
        cities:
          type: array
          items:
            type: integer
        countries:
          type: array
          items:
            type: string
        dates:
          type: object
          properties:
            start_date:
              type: string
              format: date
            end_date:
              type: string
              format: date
        filters:
          type: object
          properties:
            rating:
              type: object
              properties:
                minimum_review_score:
                  type: number
                  format: float
                minimum_review_count:
                  type: integer
        rows:
          type: integer
          example: 20
        sort:
          type: object
          properties:
            by:
              type: string
              example: most_popular
    SearchAtraccionesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AtraccionResponse'
        metadata:
          type: object
          properties:
            total_results:
              type: integer
        request_id:
          type: string
    CreateAtraccionRequest:
      type: object
      properties:
        name:
          type: string
        long_description:
          type: string
        duration:
          type: string
        price:
          $ref: '#/components/schemas/Price'
        categories:
          type: array
          items:
            type: string
        badges:
          type: array
          items:
            type: string
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        operator:
          $ref: '#/components/schemas/Operator'
        product_type:
          type: string
          enum:
            - SINGLE_TICKET
            - GUIDED_TOUR
            - PACKAGE
        includes:
          type: array
          items:
            type: string
        supported_languages:
          type: array
          items:
            type: string
        free_cancellation:
          type: boolean
    UpdateAtraccionRequest:
      type: object
      description: Update atraccion request
    AtraccionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        long_description:
          type: string
        duration:
          type: string
        price:
          $ref: '#/components/schemas/Price'
        categories:
          type: array
          items:
            type: string
        badges:
          type: array
          items:
            type: string
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        operator:
          $ref: '#/components/schemas/Operator'
        product_type:
          type: string
          enum:
            - SINGLE_TICKET
            - GUIDED_TOUR
            - PACKAGE
        includes:
          type: array
          items:
            type: string
        supported_languages:
          type: array
          items:
            type: string
        free_cancellation:
          type: boolean
        ratings:
          $ref: '#/components/schemas/Rating'
        _links:
          type: object
    PaginatedAtraccionResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AtraccionResponse'
        meta:
          type: object
          properties:
            totalItems:
              type: integer
            itemCount:
              type: integer
            itemsPerPage:
              type: integer
            totalPages:
              type: integer
            currentPage:
              type: integer
    AvailabilityResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        available_spots:
          type: integer
        times:
          type: array
          items:
            type: string
    ReservationRequest:
      type: object
      required:
        - date
        - ticket_count
        - customer_name
      properties:
        date:
          type: string
          format: date
        time:
          type: string
        ticket_count:
          type: integer
        customer_name:
          type: string
        customer_email:
          type: string
    ReservationResponse:
      type: object
      properties:
        reservation_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [CONFIRMED, PENDING, CANCELLED]
        ticket_count:
          type: integer
        total_price:
          $ref: '#/components/schemas/Price'